requirements.txt
antlr4-python3-runtime==4.9.3
contourpy==1.0.6
cycler==0.11.0
fonttools==4.38.0
hydra-core==1.2.0
importlib-resources==5.10.0
kiwisolver==1.4.4
matplotlib==3.6.2
numpy==1.23.4
omegaconf==2.2.3
opencv-python==4.6.0.66
packaging==21.3
Pillow==9.3.0
pyparsing==3.0.9
python-dateutil==2.8.2
PyYAML==6.0
six==1.16.0
zipp==3.10.0
